refactor(deepagents): move GLM-5.2 finalize/ramble middleware into harness profile
The FinalizeMiddleware (graceful budget stop) and RambleMiddleware (anti-ramble
nudge) were dcode-specific, opt-in via create_cli_agent's enable_finalize /
enable_anti_ramble flags, and only ever turned on by the harbor eval harness for
GLM-5.2. Relocate them so they live with the model they tune:

- New deepagents.profiles.harness._fireworks_glm_5p2_middleware (one module)
  holding both middlewares; the GLM-5.2 HarnessProfile attaches them via an
  extra_middleware factory, so they apply only when that profile resolves.
- Drop enable_finalize / enable_anti_ramble from dcode create_cli_agent (params,
  docstrings, append blocks) and from the harbor langgraph agent call.
- Trim the GLM-5.2 verification_discipline prompt suffix to the deliverables
  (output-coverage) clause only; drop the persistence and honest-checks clauses.
- Revert dcode default recursion_limit 2000 -> 1000 (dcode-specific config; the
  hard-turn cap now lives in the profile middleware, not dcode).

Tests: port the RambleMiddleware unit tests into libs/deepagents, add minimal
FinalizeMiddleware turn-budget tests and a profile-wiring regression test; update
the harbor agent test to assert the flags are no longer passed.
