+++ libs/code/deepagents_code/agent.py
@@ -1192,8 +1192,6 @@ def create_cli_agent(
     enable_skills: bool = True,
     enable_shell: bool = True,
     enable_interpreter: bool = False,
-    enable_finalize: bool = False,
-    enable_anti_ramble: bool = False,
     checkpointer: BaseCheckpointSaver | None = None,
     mcp_server_info: list[MCPServerInfo] | None = None,
     cwd: str | Path | None = None,
@@ -1271,12 +1269,6 @@ def create_cli_agent(
 
             Requires the `quickjs` optional extra
             (`langchain-quickjs>=0.3.1,<0.4.0`).
-        enable_finalize: Add `FinalizeMiddleware`, which nudges the agent to
-            secure a deliverable and gracefully ends the run as it nears the
-            `recursion_limit`.
-        enable_anti_ramble: Add `RambleMiddleware`, which nudges the agent to act
-            (write and run a script) and loops back to the model when a turn is
-            long prose with no tool call.
         checkpointer: Optional checkpointer for session persistence.
             When `None`, the graph is compiled without a checkpointer.
         mcp_server_info: MCP server metadata to surface in the system prompt.
@@ -1437,18 +1429,6 @@ def create_cli_agent(
     from deepagents_code.resume_state import ResumeStateMiddleware
 
     agent_middleware.append(ResumeStateMiddleware())
-
-    # Best-effort finalize: nudge + graceful stop as the run nears recursion_limit.
-    if enable_finalize:
-        from deepagents_code.finalize_middleware import FinalizeMiddleware
-
-        agent_middleware.append(FinalizeMiddleware())
-
-    # Anti-ramble: nudge + loop back when a model turn is long prose with no tool call.
-    if enable_anti_ramble:
-        from deepagents_code.ramble_middleware import RambleMiddleware
-
-        agent_middleware.append(RambleMiddleware())
 
     # Add ask_user middleware (must be early so its tool is available)
     if enable_ask_user:
